+++ packages/lit-html/src/lit-html.ts
@@ -227,6 +227,21 @@ export type TemplateResult<T extends ResultType = ResultType> = {
 
 export type SVGTemplateResult = TemplateResult<typeof SVG_RESULT>;
 
+export interface CompiledTemplateResult {
+  // This is a factory in order to make template initialization lazy
+  // and allow ShadyRenderOptions scope to be passed in.
+  _$litType$: CompiledTemplate;
+  values: unknown[];
+}
+
+export interface CompiledTemplate extends Omit<Template, 'el'> {
+  // el is overridden to be optional. We initialize it on first render
+  el?: HTMLTemplateElement;
+
+  // The prepared HTML string to create a template element from.
+  h: string;
+}
+
 /**
  * Generates a template literal tag function that returns a TemplateResult with
  * the given result type.
@@ -307,7 +322,7 @@ export const render = (
   if (part === undefined) {
     const endNode = options?.renderBefore ?? null;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (partOwnerNode as any)._$litPart$ = part = new ChildPartImpl(
+    (partOwnerNode as any)._$litPart$ = part = new ChildPart(
       container.insertBefore(createMarker(), endNode),
       endNode,
       undefined,
@@ -518,20 +533,17 @@ const getTemplateHtml = (
   ];
 };
 
-export type Template = Interface<TemplateImpl>;
-class TemplateImpl {
+export type {Template};
+class Template {
   /** @internal */
-  _$element!: HTMLTemplateElement;
+  el!: HTMLTemplateElement;
   /** @internal */
-  _parts: Array<TemplatePart> = [];
-  // Note, this is used by the `polyfill-support` module.
-  _$options?: RenderOptions;
+  parts: Array<TemplatePart> = [];
 
   constructor(
     {strings, _$litType$: type}: TemplateResult,
     options?: RenderOptions
   ) {
-    this._$options = options;
     let node: Node | null;
     let nodeIndex = 0;
     let bindingIndex = 0;
@@ -540,12 +552,12 @@ class TemplateImpl {
 
     // Create template element
     const [html, attrNames] = getTemplateHtml(strings, type);
-    this._$element = this._$createElement(html);
-    walker.currentNode = this._$element.content;
+    this.el = Template.createElement(html, options);
+    walker.currentNode = this.el.content;
 
     // Reparent SVG nodes into template root
     if (type === SVG_RESULT) {
-      const content = this._$element.content;
+      const content = this.el.content;
       const svgElement = content.firstChild!;
       svgElement.remove();
       content.append(...svgElement.childNodes);
@@ -583,25 +595,25 @@ class TemplateImpl {
                 )!;
                 const statics = value.split(marker);
                 const m = /([.?@])?(.*)/.exec(realName)!;
-                this._parts.push({
-                  _type: ATTRIBUTE_PART,
-                  _index: nodeIndex,
-                  _name: m[2],
-                  _strings: statics,
-                  _constructor:
+                this.parts.push({
+                  type: ATTRIBUTE_PART,
+                  index: nodeIndex,
+                  name: m[2],
+                  strings: statics,
+                  ctor:
                     m[1] === '.'
-                      ? PropertyPartImpl
+                      ? PropertyPart
                       : m[1] === '?'
-                      ? BooleanAttributePartImpl
+                      ? BooleanAttributePart
                       : m[1] === '@'
-                      ? EventPartImpl
-                      : AttributePartImpl,
+                      ? EventPart
+                      : AttributePart,
                 });
                 bindingIndex += statics.length - 1;
               } else {
-                this._parts.push({
-                  _type: ELEMENT_PART,
-                  _index: nodeIndex,
+                this.parts.push({
+                  type: ELEMENT_PART,
+                  index: nodeIndex,
                 });
               }
             }
@@ -626,7 +638,7 @@ class TemplateImpl {
             // normalized in some browsers (TODO: check)
             for (let i = 0; i < lastIndex; i++) {
               (node as Element).append(strings[i], createMarker());
-              this._parts.push({_type: CHILD_PART, _index: ++nodeIndex});
+              this.parts.push({type: CHILD_PART, index: ++nodeIndex});
               bindingIndex++;
             }
             (node as Element).append(strings[lastIndex] || createMarker());
@@ -636,7 +648,7 @@ class TemplateImpl {
         const data = (node as Comment).data;
         if (data === markerMatch) {
           bindingIndex++;
-          this._parts.push({_type: CHILD_PART, _index: nodeIndex});
+          this.parts.push({type: CHILD_PART, index: nodeIndex});
         } else {
           let i = -1;
           while ((i = (node as Comment).data.indexOf(marker, i + 1)) !== -1) {
@@ -644,7 +656,7 @@ class TemplateImpl {
             // The binding won't work, but subsequent bindings will
             // TODO (justinfagnani): consider whether it's even worth it to
             // make bindings in comments work
-            this._parts.push({_type: COMMENT_PART, _index: nodeIndex});
+            this.parts.push({type: COMMENT_PART, index: nodeIndex});
             bindingIndex++;
             // Move to the end of the match
             i += marker.length - 1;
@@ -656,10 +668,10 @@ class TemplateImpl {
   }
 
   // Overridden via `litHtmlPlatformSupport` to provide platform support.
-  _$createElement(html: string) {
-    const template = d.createElement('template');
-    template.innerHTML = html;
-    return template;
+  static createElement(html: string, _options?: RenderOptions) {
+    const el = d.createElement('template');
+    el.innerHTML = html;
+    return el;
   }
 }
 
@@ -671,8 +683,8 @@ export interface Disconnectable {
 function resolveDirective(
   part: ChildPart | AttributePart | ElementPart,
   value: unknown,
-  _$parent: DirectiveParent = part,
-  _$attributeIndex?: number
+  parent: DirectiveParent = part,
+  attributeIndex?: number
 ): unknown {
   // Bail early if the value is explicitly noChange. Note, this means any
   // nested directive is still attached and is not run.
@@ -680,29 +692,26 @@ function resolveDirective(
     return value;
   }
   let currentDirective =
-    _$attributeIndex !== undefined
-      ? (_$parent as AttributePart).__directives?.[_$attributeIndex]
-      : (_$parent as ChildPart | ElementPart | Directive).__directive;
+    attributeIndex !== undefined
+      ? (parent as AttributePart).__directives?.[attributeIndex]
+      : (parent as ChildPart | ElementPart | Directive).__directive;
   const nextDirectiveConstructor = isPrimitive(value)
     ? undefined
     : (value as DirectiveResult)._$litDirective$;
   if (currentDirective?.constructor !== nextDirectiveConstructor) {
     currentDirective?._$setDirectiveConnected?.(false);
-    currentDirective =
-      nextDirectiveConstructor === undefined
-        ? undefined
-        : new nextDirectiveConstructor({
-            ...part,
-            _$part: part,
-            _$parent,
-            _$attributeIndex,
-          } as PartInfo);
-    if (_$attributeIndex !== undefined) {
-      ((_$parent as AttributePart).__directives ??= [])[
-        _$attributeIndex
+    if (nextDirectiveConstructor === undefined) {
+      currentDirective = undefined;
+    } else {
+      currentDirective = new nextDirectiveConstructor(part as PartInfo);
+      currentDirective._$initialize(part, parent, attributeIndex);
+    }
+    if (attributeIndex !== undefined) {
+      ((parent as AttributePart).__directives ??= [])[
+        attributeIndex
       ] = currentDirective;
     } else {
-      (_$parent as ChildPart | Directive).__directive = currentDirective;
+      (parent as ChildPart | Directive).__directive = currentDirective;
     }
   }
   if (currentDirective !== undefined) {
@@ -710,7 +719,7 @@ function resolveDirective(
       part,
       currentDirective._$resolve(part, (value as DirectiveResult).values),
       currentDirective,
-      _$attributeIndex
+      attributeIndex
     );
   }
   return value;
@@ -740,8 +749,8 @@ class TemplateInstance {
   // DocumentFragment and we don't want to hold onto it with an instance field.
   _clone(options: RenderOptions | undefined) {
     const {
-      _$element: {content},
-      _parts: parts,
+      el: {content},
+      parts: parts,
     } = this._$template;
     const fragment = (options?.creationScope ?? d).importNode(content, true);
     walker.currentNode = fragment;
@@ -752,30 +761,30 @@ class TemplateInstance {
     let templatePart = parts[0];
 
     while (templatePart !== undefined && node !== null) {
-      if (nodeIndex === templatePart._index) {
+      if (nodeIndex === templatePart.index) {
         let part: Part | undefined;
-        if (templatePart._type === CHILD_PART) {
-          part = new ChildPartImpl(
+        if (templatePart.type === CHILD_PART) {
+          part = new ChildPart(
             node as HTMLElement,
             node.nextSibling,
             this,
             options
           );
-        } else if (templatePart._type === ATTRIBUTE_PART) {
-          part = new templatePart._constructor(
+        } else if (templatePart.type === ATTRIBUTE_PART) {
+          part = new templatePart.ctor(
             node as HTMLElement,
-            templatePart._name,
-            templatePart._strings,
+            templatePart.name,
+            templatePart.strings,
             this,
             options
           );
-        } else if (templatePart._type === ELEMENT_PART) {
-          part = new ElementPartImpl(node as HTMLElement, this, options);
+        } else if (templatePart.type === ELEMENT_PART) {
+          part = new ElementPart(node as HTMLElement, this, options);
         }
         this._parts.push(part);
         templatePart = parts[++partIndex];
       }
-      if (templatePart !== undefined && nodeIndex !== templatePart._index) {
+      if (templatePart !== undefined && nodeIndex !== templatePart.index) {
         node = walker.nextNode();
         nodeIndex++;
       }
@@ -805,42 +814,26 @@ class TemplateInstance {
 /*
  * Parts
  */
-type AttributePartConstructor = {
-  new (
-    element: HTMLElement,
-    name: string,
-    strings: ReadonlyArray<string>,
-    parent: Disconnectable | undefined,
-    options: RenderOptions | undefined
-  ): AttributePart;
-};
 type AttributeTemplatePart = {
-  readonly _type: typeof ATTRIBUTE_PART;
-  readonly _index: number;
-  readonly _name: string;
+  readonly type: typeof ATTRIBUTE_PART;
+  readonly index: number;
+  readonly name: string;
   /** @internal */
-  readonly _constructor: AttributePartConstructor;
+  readonly ctor: typeof AttributePart;
   /** @internal */
-  readonly _strings: ReadonlyArray<string>;
+  readonly strings: ReadonlyArray<string>;
 };
 type NodeTemplatePart = {
-  readonly _type: typeof CHILD_PART;
-  readonly _index: number;
-};
-type ElementPartConstructor = {
-  new (
-    element: HTMLElement,
-    parent: Disconnectable | undefined,
-    options: RenderOptions | undefined
-  ): ElementPart;
+  readonly type: typeof CHILD_PART;
+  readonly index: number;
 };
 type ElementTemplatePart = {
-  readonly _type: typeof ELEMENT_PART;
-  readonly _index: number;
+  readonly type: typeof ELEMENT_PART;
+  readonly index: number;
 };
 type CommentTemplatePart = {
-  readonly _type: typeof COMMENT_PART;
-  readonly _index: number;
+  readonly type: typeof COMMENT_PART;
+  readonly index: number;
 };
 
 /**
@@ -862,10 +855,8 @@ export type Part =
   | ElementPart
   | EventPart;
 
-type Interface<T> = {[P in keyof T]: T[P]};
-
-export type ChildPart = Interface<ChildPartImpl>;
-class ChildPartImpl {
+export type {ChildPart};
+class ChildPart {
   readonly type = CHILD_PART;
   readonly options: RenderOptions | undefined;
   _$committedValue: unknown;
@@ -1046,13 +1037,28 @@ class ChildPartImpl {
     this._$committedValue = value;
   }
 
-  private _commitTemplateResult(result: TemplateResult): void {
-    const {values, strings} = result;
-    const template = this._$getTemplate(strings, result);
+  private _commitTemplateResult(
+    result: TemplateResult | CompiledTemplateResult
+  ): void {
+    const {values, _$litType$} = result;
+    // If $litType$ is a number, result is a plain TemplateResult and we get
+    // the template from the template cache. If not, result is a
+    // CompiledTemplateResult and _$litType$ is a CompiledTemplate and we need
+    // to create the <template> element the first time we see it.
+    const template: Template | CompiledTemplate =
+      typeof _$litType$ === 'number'
+        ? this._$getTemplate(result as TemplateResult)
+        : (_$litType$.el === undefined &&
+            (_$litType$.el = Template.createElement(
+              _$litType$.h,
+              this.options
+            )),
+          _$litType$);
+
     if ((this._$committedValue as TemplateInstance)?._$template === template) {
       (this._$committedValue as TemplateInstance)._update(values);
     } else {
-      const instance = new TemplateInstance(template!, this);
+      const instance = new TemplateInstance(template as Template, this);
       const fragment = instance._clone(this.options);
       instance._update(values);
       this._commitNode(fragment);
@@ -1062,10 +1068,10 @@ class ChildPartImpl {
 
   // Overridden via `litHtmlPlatformSupport` to provide platform support.
   /** @internal */
-  _$getTemplate(strings: TemplateStringsArray, result: TemplateResult) {
-    let template = templateCache.get(strings);
+  _$getTemplate(result: TemplateResult) {
+    let template = templateCache.get(result.strings);
     if (template === undefined) {
-      templateCache.set(strings, (template = new TemplateImpl(result)));
+      templateCache.set(result.strings, (template = new Template(result)));
     }
     return template;
   }
@@ -1088,9 +1094,9 @@ class ChildPartImpl {
 
     // Lets us keep track of how many items we stamped so we can clear leftover
     // items from a previous render
-    const itemParts = this._$committedValue as ChildPartImpl[];
+    const itemParts = this._$committedValue as ChildPart[];
     let partIndex = 0;
-    let itemPart: ChildPartImpl | undefined;
+    let itemPart: ChildPart | undefined;
 
     for (const item of value) {
       if (partIndex === itemParts.length) {
@@ -1099,7 +1105,7 @@ class ChildPartImpl {
         // instead of sharing parts between nodes
         // https://github.com/Polymer/lit-html/issues/1266
         itemParts.push(
-          (itemPart = new ChildPartImpl(
+          (itemPart = new ChildPart(
             this._insert(createMarker()),
             this._insert(createMarker()),
             this,
@@ -1149,8 +1155,8 @@ class ChildPartImpl {
   }
 }
 
-export type AttributePart = Interface<AttributePartImpl>;
-class AttributePartImpl {
+export type {AttributePart};
+class AttributePart {
   readonly type = ATTRIBUTE_PART as
     | typeof ATTRIBUTE_PART
     | typeof PROPERTY_PART
@@ -1304,8 +1310,8 @@ class AttributePartImpl {
   }
 }
 
-export type PropertyPart = Interface<PropertyPartImpl>;
-class PropertyPartImpl extends AttributePartImpl {
+export type {PropertyPart};
+class PropertyPart extends AttributePart {
   readonly type = PROPERTY_PART;
 
   /** @internal */
@@ -1325,8 +1331,8 @@ class PropertyPartImpl extends AttributePartImpl {
   }
 }
 
-export type BooleanAttributePart = Interface<BooleanAttributePartImpl>;
-class BooleanAttributePartImpl extends AttributePartImpl {
+export type {BooleanAttributePart};
+class BooleanAttributePart extends AttributePart {
   readonly type = BOOLEAN_ATTRIBUTE_PART;
 
   /** @internal */
@@ -1353,8 +1359,8 @@ type EventListenerWithOptions = EventListenerOrEventListenerObject &
  * Because event options are passed when adding listeners, we must take case
  * to add and remove the part as a listener when the event options change.
  */
-export type EventPart = Interface<EventPartImpl>;
-class EventPartImpl extends AttributePartImpl {
+export type {EventPart};
+class EventPart extends AttributePart {
   readonly type = EVENT_PART;
 
   // EventPart does not use the base _$setValue/_resolveValue implementation
@@ -1407,7 +1413,7 @@ class EventPartImpl extends AttributePartImpl {
 
   handleEvent(event: Event) {
     if (typeof this._$committedValue === 'function') {
-      // TODO (justinfagnani): do we need to default to this._$element?
+      // TODO (justinfagnani): do we need to default to this.element?
       // It'll always be the same as `e.currentTarget`.
       this._$committedValue.call(this.options?.host ?? this.element, event);
     } else {
@@ -1416,8 +1422,8 @@ class EventPartImpl extends AttributePartImpl {
   }
 }
 
-export type ElementPart = Interface<ElementPartImpl>;
-class ElementPartImpl {
+export type {ElementPart};
+class ElementPart {
   readonly type = ELEMENT_PART;
 
   /** @internal */
@@ -1485,20 +1491,20 @@ export const _Σ = {
   _isIterable: isIterable,
   _resolveDirective: resolveDirective,
   // Used in tests and private-ssr-support
-  _ChildPart: ChildPartImpl,
-  _AttributePart: AttributePartImpl as AttributePartConstructor,
-  _BooleanAttributePart: BooleanAttributePartImpl as AttributePartConstructor,
-  _EventPart: EventPartImpl as AttributePartConstructor,
-  _PropertyPart: PropertyPartImpl as AttributePartConstructor,
-  _ElementPart: ElementPartImpl as ElementPartConstructor,
+  _ChildPart: ChildPart,
+  _AttributePart: AttributePart,
+  _BooleanAttributePart: BooleanAttributePart,
+  _EventPart: EventPart,
+  _PropertyPart: PropertyPart,
+  _ElementPart: ElementPart,
 };
 
 // Apply polyfills if available
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-(globalThis as any)['litHtmlPlatformSupport']?.(TemplateImpl, ChildPartImpl);
+(globalThis as any)['litHtmlPlatformSupport']?.(Template, ChildPart);
 
 // IMPORTANT: do not change the property name or the assignment expression.
 // This line will be used in regexes to search for lit-html usage.
 // TODO(justinfagnani): inject version number at build time
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-((globalThis as any)['litHtmlVersions'] ??= []).push('2.0.0-pre.6');
+((globalThis as any)['litHtmlVersions'] ??= []).push('2.0.0-pre.7');