+++ packages/localize-tools/CHANGELOG.md
@@ -7,6 +7,62 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## Unreleased -->
 
+## [0.2.1] - 2021-04-02
+
+### Changed
+
+- XLIFF file headers have been simplified to:
+
+```xml
+<xliff version="1.2" xmlns="urn:oasis:names:tc:xliff:document:1.2">
+```
+
+## [0.2.0] - 2021-03-30
+
+### Changed
+
+- **[BREAKING]** Description comments (`// msgdesc:`) have been removed in favor
+  of the `desc` option.
+
+Before:
+
+```js
+// msgdesc: Home page
+class HomePage {
+  hello() {
+    // msgdesc: Greeting to Earth
+    return msg(html`Hello World`);
+  }
+  goodbye() {
+    // msgdesc: Farewell to Earth
+    return msg(html`Goodbye World`);
+  }
+}
+```
+
+After:
+
+```js
+class HomePage {
+  hello() {
+    return msg(html`Hello World`, {
+      desc: 'Home page / Greeting to Earth',
+    });
+  }
+  goodbye() {
+    return msg(html`Goodbye World`, {
+      desc: 'Home page / Farewell to Earth',
+    });
+  }
+}
+```
+
+## [0.1.1] - 2021-03-30
+
+### Changed
+
+- Bumped dependency versions for `xmldom` and `@lit/localize`
+
 ## [0.1.0] - 2021-03-24
 
 ### Changed